# src/main/resources/v1/eric-oss-performance-monitoring-enabler-openapi.yaml
...
openapi: 3.0.2
servers:
  - url : http://localhost/performance-monitoring-enabler/v1alpha3
info:
  x-api-id: 9a23a7fc-5aa5-469f-a234-9b47f3c3955e
  x-audience: external-partner
  description: |
    Performance Monitoring Enabler (PME) provides the capability to monitor 
    KPIs in the network and determine if they fall outside a certain threshold. 
    The KPIs and the threshold information is defined in a configuration, 
    while a session allows the client to control how long these KPIs are 
    monitored for.  A configuration can be shared among several sessions.
  version: 0.0.3
  title: Performance Monitoring Enabler API
  license:
    name: COPYRIGHT Ericsson 2024
    url: https://www.ericsson.com/en/legal
  termsOfService: https://www.ericsson.com/en/legal
  contact:
    email: pdlteamvad@pdl.internal.ericsson.com
...
paths:
  /sessions:
    get:
      tags:
        - sessions
      summary: Get the list of PME Monitoring Sessions
      description: Gets all the PME Monitoring Session for the given rApp Client ID
      operationId: getSessions
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: "#/components/parameters/clientId"  # @TODO clientId is a temporary solution and will be removed before first release
        - name: sessionReference
          in: query
          description: A client reference that is unique to the session.
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "503":
          $ref: "#/components/responses/503"
        default:
          $ref: "#/components/responses/500"
    post:
      tags:
        - sessions
      summary: Create an PME Monitoring Session
      description: Creates an PME Monitoring Session for the given rApp Client ID
      operationId: createSession
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/contentTypeInHeader'
        - $ref: "#/components/parameters/clientId"  # @TODO clientId is a temporary solution and will be removed before first release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionRequest'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/400"
        "409":
          $ref: "#/components/responses/409"
        "415":
          $ref: "#/components/responses/415"
        "503":
          $ref: "#/components/responses/503"
        default:
          $ref: "#/components/responses/500"
  /sessions/{sessionId}:
    get:
      tags:
        - sessions
      summary: Get an PME Monitoring Session
      description: Get an PME Monitoring Session for the given rApp Client ID and Session ID
      operationId: getSessionById
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: "#/components/parameters/clientId"  # @TODO clientId is a temporary solution and will be removed before first release
        - $ref: "#/components/parameters/sessionId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          $ref: "#/components/responses/404"
        "503":
          $ref: "#/components/responses/503"
        default:
          $ref: "#/components/responses/500"
  /sessions/{sessionId}/status:
    put:
      tags:
        - sessions
      summary: Stop an PME Monitoring Session
      description: Stop an PME Monitoring Session for the given rApp Client ID and Session ID
      operationId: stopSessionById
      parameters:
        - $ref: '#/components/parameters/acceptInHeader'
        - $ref: '#/components/parameters/contentTypeInHeader'
        - $ref: "#/components/parameters/clientId"  # @TODO clientId is a temporary solution and will be removed before first release
        - $ref: "#/components/parameters/sessionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionStopRequest'
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStopRequest"
        "400":
          $ref: "#/components/responses/400"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "415":
          $ref: "#/components/responses/415"
        "503":
          $ref: "#/components/responses/503"
        default:
          $ref: "#/components/responses/500"
...
components:
...
  parameters:
    clientId:
      name: clientId
      in: query
      description: ID of the Client who owns the session
      required: true
      schema:
        type: string
        minLength: 4
        maxLength: 64
        pattern: "^[\\w-]+$"
    sessionId:
      name: sessionId
      in: path
      description: ID of session
      required: true
      schema:
        type: string
...
    acceptInHeader:
      name: Accept
      in: header
      required: false
      schema:
        type: string
        example: application/json
        default: application/json
    contentTypeInHeader:
      name: Content-Type
      in: header
      required: false
      schema:
        type: string
        example: application/json
        default: application/json

  # Common error responses
  responses:
    "400":
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorDetails"
          example: {
            "title": "Validation Failed",
            "status": 400,
            "detail": "Failed to read request"
          }
    "404":
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorDetails"
          example: {
            "title": "Provide a valid Configuration ID",
            "status": 404,
            "detail": "Configuration with provided ID does not exist"
          }
    "409":
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorDetails"
          example: {
            "title": "Only Configurations not associated with a Session can be updated",
            "status": 409,
            "detail": "Configuration with provided ID is used in a Session"
          }
    "415":
      description: Unsupported Media Type
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorDetails"
          example: {
            "title": "Validation Failed",
            "status": 415,
            "detail": "Content-Type 'text/plain;charset=ISO-8859-1' is not supported"
          }
    "503":
      description: Service Unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorDetails"
          example: {
            "title": "Service is not ready",
            "status": 503,
            "detail": "Start up operations have started"
          }
    "500":
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ErrorDetails"
          example: {
            "title": "Internal Server Error",
            "status": 500,
            "detail": "Failed to persist change to the database"
          }
  schemas:
    SessionRequest:
      description: The session sent by the client.
      type: object
      required:
        - sessionReference
        - pmeConfigId
      properties:
        sessionReference:
          description: A client reference that is unique to the session.
          type: string
          minLength: 4
          maxLength: 64
          pattern: "^[\\w-_]+$"
        duration:
          description: The number of hours for which the session will be active and monitoring will be performed.
          type: integer
          maximum: 24
          minimum: 1
          default: 18
          format: int32
        pmeConfigId:
          description: The id of the configuration to be used by this session.
          type: string
    SessionStopRequest:
      description: The payload to stop an PME Session
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - STOPPED
    Session:
      description: The created session.
      type: object
      required:
        - id
        - sessionReference
        - duration
        - pmeConfigId
        - monitoringObjectTopicName
        - verdictTopicName
        - status
        - createdAt
      properties:
        id:
          type: string
          minLength: 4
          maxLength: 255
          pattern: "^[\\w-_]+$"
        sessionReference:
          type: string
          minLength: 4
          maxLength: 64
          pattern: "^[\\w-_]+$"
        duration:
          type: integer
          maximum: 24
          minimum: 1
          format: int32
        pmeConfigId:
          type: string
        monitoringObjectTopicName:
          type: string
        verdictTopicName:
          type: string
        status:
          type: string
          enum:
            - CREATED
            - STARTED
            - STOPPED
            - FINISHED
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        stoppedAt:
          type: string
          format: date-time
          nullable: true
...
    ErrorDetails:
      description: Provides additional information in an error response.
      type: object
      properties:
        title:
          type: string
          description: Short, human-readable summary of the problem.
        status:
          type: integer
          description: HTTP status code.
        detail:
          type: string
          description: Optional human-readable explanation of the problem. if present, ought to focus on helping the client correct the problem, rather than giving debugging information.